Kafka - further metrics & debug

+++ open-metadata-implementation/adapters/open-connectors/event-bus-connectors/open-metadata-topic-connectors/kafka-open-metadata-topic-connector/src/main/java/org/odpi/openmetadata/adapters/eventbus/topic/kafka/KafkaOpenMetadataEventProducer.java
@@ -40,8 +40,8 @@ public class KafkaOpenMetadataEventProducer implements Runnable {
     private long kafkaSendAttemptCount = 0;
     private long messagePublishRequestCount = 0;
     private long inmemoryPutMessageCount = 0;
-    private int kafkaSendFailCount = 0;
-    private int messageFailedSendCount = 0;
+    private long kafkaSendFailCount = 0;
+    private long messageFailedSendCount = 0;
 
 
     /**
